First iteration, unit tests pass.
